(maint) Remove udev persistent net naming
By default on rhel/centos, detected ethernet interfaces are added to the
/etc/udev/rules.d/70-persistent-net.rules file to ensure consistent
device addressing. This breaks with VMs, where devices and addressing
needs to be fluid as they move around.

This removes the file so that no static device naming will be enforced.
Previously, we were doing this in the kickstart file, but now that we are
using vmpooler VMs, we need to remove it in the configuration stage.