Preseed classroom module on first boot
This will put all the required modules in both code-staging and in
codedir, with the proper permissions set. This means that it is useful
with and without code manager turned on.

It will also generate the deployer user and configure an access token.

This is not my favorite approach, but this is a finnicky tool to setup
and I cannot find a better way.

COURSES-2217 #time 12h