This adds a script to help resolve module dependencies
This will just force remove all dependencies of a module. This means
that when you try to install it again, it will only use the dependencies
described in that module, rather than trying to reconcile it with the
currently installed module versions.

WARNING: This will **remove** modules from your modulepath.

Note: I'm not entirely certain this is really a good idea. I'd rather
that people were able and willing to debug module dependencies on their
own rather than running this nuke-from-orbit script. My worry is that if
this does resolve their issues that they'll never get around to telling
us so that we can fix it, and that they won't understand that it's
actually removing modules that they might need to reinstall.