Support offline module installation
This extends the auto-installation logic to also work when modules cannot be installed from the forge. This will cover offline classes, and when upstream modules break dependency resolution.

The one concern here, is that it's a silent failure. It's likely that we won't get reports of invalid dependencies from the field. That's *probably* ok, since we'll be testing dependency resolution each time the VM is built.