Cache the courseware gem
This adds the ability to cache gems by source URLs and caches the
courseware gem. This will rely on us keeping the S3 copy as up-to-date
as we need it to be.

Next sprint after we're certain that all instructors have updated their
VM, convert the classroom pdf stack manifest to use this.

COURSES-2001 #resolved #time 1h