Make yum cache idempotent & more readable
This has the primary goal of making the package list more readable, and
of recording which packages are cached intentionally vs as a dependency.
It also lets us make the class idempotent.

To force a repo rebuild and check for package updates, you could delete
the `repotrack.pkg` file before running. Or we could just turn off
`refreshonly => true` and give up the idempotency.