Remove conflicting hiera data
This was copy-pasta into the bootstrap for some reason. Not only do we
not need it (the important setting is managed by the classroom module
when needed), but it causes flap when it tries to own hiera files and
then the classroom module tries to own the same files.

TRAINTECH-1396 #resolved #time 90m