(TRAINTECH-1409) Add domains for hiera course to ssh_config
The /etc/ssh_config file on the Learning VM contains a section that
bypasses host key checking for container nodes. Prior to this commit, it
only does this for hosts following the *.puppet.vm pattern. This commit
adds *.auroch.vm and *.beauvine.vm domains used in the Hiera quest to
this pattern.