This should persist the http redirect after install
I am pretty sure that what's happening here is that the installer
configures PE with the HTTP redirect disabled. But then it doesn't
persist the setting, so it gets turned back on (the default) as soon as
the first Puppet run happens.

This should correct that, though this file will be wiped out as soon as
the first Puppet code deploy happens. By then, we'll have reconfigured
and it won't matter though.

TRAINTECH-1564 #resolved #time 30m