Manage codebase parts all in one place.
This doesn't drop the `common.yaml` before install, but I realized that
it shouldn't matter since the install is running off the `pe.conf`
template. This will exist prior to the first PE agent run, and then when
the agent is classified and code deployment happens, it will all be
replaced anyway. (There's a small race condition in that, so there might
still be a Puppet run that flaps this setting, but that should go away
once we get the hiera configs all updated to match current best
practices.)

TRAINTECH-1564 #time 2h